refactor:  separate HttpProcess and create WebProcessor
